step1b: NaN-proof training + trustworthy eval + one-shot full pipeline job
- train loop: skip opt.step()/EMA on non-finite grad norm (a single bf16 spike
  made clip_grad_norm_ rescale 0*inf=NaN and opt.step() poisoned all weights +
  AdamW moments + EMA permanently; job 312503 died ~ep75, NaN to ep150).
- model_tokens: compute the 1/||reg||-gradient cosine loss in fp32 (likely spike source).
- train eval: stratify the decode sample ACROSS subjects instead of first decode_n
  (= subject-1 only) — 16-image CLIP_2way had ~0.1 SE, making the metric noise and
  best-ckpt selection random.
- slurm/step1b_full_a100.sbatch: train then auto-run a full-test-set eval in one job.

+++ brainflow/step1/model_tokens.py
@@ -213,7 +213,9 @@ class TokenStep1Model(nn.Module):
 
         reg = self.reg_head(brain)
         loss_reg = F.mse_loss(reg, target_std)
-        loss_cos = 1.0 - F.cosine_similarity(reg, target_std, dim=-1).mean()
+        # fp32 cosine: its gradient scales as 1/‖reg‖ and is the most likely bf16
+        # overflow source; compute it in fp32 to keep the spike from ever starting.
+        loss_cos = 1.0 - F.cosine_similarity(reg.float(), target_std.float(), dim=-1).mean()
 
         x1 = target_std
         x0 = torch.randn_like(x1)
